// src/interfaces/enums/cultivos/tipo-cultivo.enums.ts
export enum TipoCultivoEnum {
  MAIZ = "Maíz",
  FRIJOL = "Frijol",
  ARROZ = "Arroz",
  SORGO = "Sorgo",
  CAFE = "Café",
  PAPA = "Papa",
  TOMATE = "Tomate",
  CEBOLLA = "Cebolla",
  AJO = "Ajo",
  YUCA = "Yuca",
  HORTALIZAS = "Hortalizas",
  FRUTAS = "Frutas",
  OTROS = "Otros",
}

export enum TipoSistemaRiego {
  GOTEO = "Goteo",
  ASPERSION = "Aspersión",
  MICROASPERSION = "Microaspersión",
  INUNDACION = "Inundación",
  GRAVEDAD = "Gravedad",
  PIVOTE_CENTRAL = "Pivote Central",
  MANUAL = "Manual",
  SECANO = "Secano",
  HIDROPONICO = "Hidropónico",
  OTRO = "Otro",
}

export enum MetodoSiembra {
  DIRECTA = "Siembra Directa",
  TRANSPLANTE = "Trasplante",
  AL_VOLEO = "Al voleo",
  SURCOS = "Surcos",
  MANUAL = "Manual",
  MECANIZADA = "Mecanizada",
  ALMACIGO = "Almácigo",
  HIDROPONICA = "Hidropónica",
  INVERNADERO = "Invernadero",
  OTRO = "Otro",
}

export enum TipoSuelo {
  ARCILLOSO = "Arcilloso",
  FRANCO = "Franco",
  ARENOSO = "Arenoso",
  LIMOSO = "Limoso",
  ORGANICO = "Orgánico",
  OTRO = "Otro",
}